move builder to new package
